gen_ai_01_news_article/requirements.txt
numpy==1.26.4
streamlit==1.49.1
langchain==0.3.27
langchain-community==0.3.29
langchain-huggingface==0.3.1
langchain-groq==0.3.8
chromadb==1.1.0
sentence-transformers==2.7.0
torch==2.8.0
huggingface-hub==0.35.0
beautifulsoup4==4.13.5
lxml==6.0.1
requests==2.32.5
python-dotenv==1.1.1
pydantic==2.11.9
